requirements/base.txt
asgiref==3.11.1
attrs==25.4.0
colorama==0.4.6
Django==6.0.2
django-environ==0.13.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.1.1
factory_boy==3.3.3
Faker==40.4.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-django==4.12.0
pytest-factoryboy==2.8.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
